Restructure servicios component and update navigation
- Create new servicios index component in medicos section
- Remove legacy servicios-tab component from website/clinica
- Update hospital-nav component to reflect structural changes
- Reorganize services-related components for better maintainability
